requirements.txt
requests
alive-progress
Pillow